python-wrapper/pyproject.toml
[tool.poetry]
name = "line-solver"
version = "3.0.2"
description = "Queueing theory algorithms in Python"
authors = [
    "Giuliano Casale <g.casale@imperial.ac.uk>"
]
license = "BSD-3"
readme = "README.md"
homepage = "https://line-solver.sourceforge.net"
packages = [{ include = "line_solver" }]

include = [
    "line_solver/jline.jar",
    "tests/**",
    "examples/**",
    "gettingstarted/**",
    "AUTHORS",
    "LICENSE",
    "README.md",
    "mm1.py",
    "mm1.ipynb",
    "requirements.txt"
]

[tool.poetry.dependencies]
python = ">=3.11"
numpy = "^2.3.3"
matplotlib = "^3.10.6"
jpype1 = "^1.6.0"
pandas = "^2.3.2"
scipy = "^1.16.2"
enum-tools = "^0.13.0"
websockets = "^12.0"
simpy = "^4.0.0"

# Optional dependencies for native backend
torch = { version = ">=2.0.0", optional = true }
numba = { version = ">=0.57.0", optional = true }

[tool.poetry.extras]
native = ["torch", "numba"]
gpu = ["torch"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
nbformat = "^5.10.4"
nbconvert = "^7.16.6"
ipykernel = "^6.30.1"

[tool.poetry.scripts]
line = "line_solver.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


